Move qualifiedTitle to Reporter utilitites, export runBenchmark from runBenchmark
